start fixing old code builds and renaming package paths
