+++ package.json
@@ -28,19 +28,19 @@
     "lint": "eslint src/app/pixi/**/*.js",
     "postinstall": "./bin/custom-build.sh",
     "prebuild": "mkdir -p public/javascripts/browserify",
-    "build:all": "npm run build && npm run build:cordova && npm run yuidoc",
+    "build:old": "npm run browserify:canvas && npm run uglify:canvas",
     "build:webgl": "npm run browserify:webgl",
     "build:custom": "npm run browserify:gobbapeas",
     "getAssets": "curl -O $THRUST_ASSETS_PACK && tar -xzf assets.tar.gz",
     "debug:android": "echo 'Ensure Android device is connected, and has USB debugging enabled' && adb forward tcp:9222 localabstract:chrome_devtools_remote",
     "deploy:skipTests": "npm run heroku",
     "preversion": "npm test -s",
-    "browserify:canvas": "echo 'browserifying canvas build' && browserify --version && browserify -e src/app/main-canvas.js -o public/javascripts/browserify/thrust-engine-canvas.js -d",
-    "browserify:mobile": "echo 'browserifying mobile build' && browserify --version && browserify -e src/app/main-mobile.js -o public/javascripts/browserify/thrust-engine-mobile.js -d",
-    "browserify:webgl": "echo 'browserifying webgl build' && browserify --version && browserify -e src/app/main-webgl.js -o public/javascripts/browserify/thrust-engine-webgl.js -d",
-    "browserify:main": "echo 'browserifying auto build' && browserify --version && browserify -e src/app/main-auto.js -o public/javascripts/browserify/thrust-engine-auto.js -d",
-    "browserify:cordova": "echo 'browserifying cordova build' && browserify --version && browserify -e src/app/main-cordova.js -o public/javascripts/browserify/thrust-engine-cordova.js -d",
-    "browserify:gobbapeas": "echo 'browserifying gobbapeas build' && browserify --version && browserify -e src/app/main-gobbapeas.js -o public/javascripts/browserify/thrust-engine-gobbapeas.js -d",
+    "browserify:canvas": "echo 'browserifying canvas build' && browserify --version && browserify -e src/old/main-canvas.js -o public/javascripts/browserify/thrust-engine-canvas.js -d",
+    "browserify:mobile": "echo 'browserifying mobile build' && browserify --version && browserify -e src/old/main-mobile.js -o public/javascripts/browserify/thrust-engine-mobile.js -d",
+    "browserify:webgl": "echo 'browserifying webgl build' && browserify --version && browserify -e src/old/main-webgl.js -o public/javascripts/browserify/thrust-engine-webgl.js -d",
+    "browserify:main": "echo 'browserifying auto build' && browserify --version && browserify -e src/old/main-auto.js -o public/javascripts/browserify/thrust-engine-auto.js -d",
+    "browserify:cordova": "echo 'browserifying cordova build' && browserify --version && browserify -e src/old/main-cordova.js -o public/javascripts/browserify/thrust-engine-cordova.js -d",
+    "browserify:gobbapeas": "echo 'browserifying gobbapeas build' && browserify --version && browserify -e src/old/main-gobbapeas.js -o public/javascripts/browserify/thrust-engine-gobbapeas.js -d",
     "uglify:canvas": "echo 'Uglifying bundle' && uglifyjs public/javascripts/browserify/thrust-engine-canvas.js -o public/javascripts/browserify/thrust-engine-canvas.min.js --source-map public/javascripts/browserify/thrust-engine.min.js.map --screw-ie8",
     "changelog": "echo 'updating changelog' && github_changelog_generator",
     "tweet:deploy": "node ./bin/tweet-deploy",
