+++ public/assets/physics-new/level-1-phyics.json
@@ -1,0 +1,20 @@
+{ 
+    
+    "level-1": [
+        {
+            "shape": [ 139,9, 227.97203063964844,59.580421447753906, 0,140, 0,9 ]
+        },
+        {
+            "shape": [ 767,140, 514.335693359375,94.89510345458984, 587.7622680664062,50.139862060546875, 659.7902221679688,10.629364013671875, 767.1328735351562,10.27972412109375 ]
+        },
+        {
+            "shape": [ 587.7622680664062,50.139862060546875, 514.335693359375,94.89510345458984, 514,50 ]
+        },
+        {
+            "shape": [ 386.01397705078125,95.94405746459961, 0,140, 227.97203063964844,59.580421447753906, 318.5314636230469,59.580421447753906 ]
+        },
+        {
+            "shape": [ 386.01397705078125,95.94405746459961, 514.335693359375,94.89510345458984, 767,140, 0,140 ]
+        }
+    ]
+}
