+++ src/old/states/launcher.js
@@ -1,10 +1,6 @@
 'use strict';
 
 var properties = require('../properties');
-var PIXI = global.PIXI;
-
-var load = require('../pixi/states/load');
-var gameLoop = require('../pixi/rendering/game-loop');
 
 /**
  * Game entry point
