+++ src/old/states/play.js
@@ -1,29 +1,29 @@
 'use strict';
 
 var properties = require('../properties');
-var Collisions = require('.././Collisions');
-var Groups = require('.././Groups');
-var ui = require('./index');
-var Player = require('.././Player');
-var Fuel = require('.././Fuel');
-var Limpet = require('.././Limpet');
-var Switch = require('.././GateSwitch');
-var Orb = require('.././Orb');
-var Background = require('.././Background');
-var TractorBeam = require('.././TractorBeam');
+var Collisions = require('../environment/Collisions');
+var Groups = require('../environment/Groups');
+var ui = require('../ui');
+var Player = require('../actors/Player');
+var Fuel = require('../actors/Fuel');
+var Limpet = require('../actors/Limpet');
+var Switch = require('../actors/GateSwitch');
+var Orb = require('../actors/Orb');
+var Background = require('../actors/Background');
+var TractorBeam = require('../actors/TractorBeam');
 var _ = require('lodash');
-var particles = require('.././particles/manager');
-var levelManager = require('.././level-manager');
-var PowerStation = require('.././PowerStation');
-var PhysicsActor = require('.././PhysicsActor');
-var gameState = require('.././game-state');
-var sound = require('./sound');
-var droneManager = require('.././drone-manager');
-var Stopwatch = require('./Stopwatch');
+var particles = require('../environment/particles/manager');
+var levelManager = require('../data/level-manager');
+var PowerStation = require('../actors/PowerStation');
+var PhysicsActor = require('../actors/PhysicsActor');
+var gameState = require('../data/game-state');
+var sound = require('../utils/sound');
+var droneManager = require('../actors/drone-manager');
+var Stopwatch = require('../ui/Stopwatch');
 var TimelineMax = global.TimelineMax;
-var features = require('./features');
-var MapAtlas = require('.././levels/MapAtlas');
-var options = require('.././options-model');
+var features = require('../utils/features');
+var MapAtlas = require('../environment/levels/MapAtlas');
+var options = require('../data/options-model');
 var StatusBar = global.StatusBar;
 
 /**
